+++ source/openta/ta_platform_layer.c
@@ -452,7 +452,7 @@ int ta_main_loop(ta_game* pGame)
         if (PeekMessageA(&msg, NULL, 0, 0, PM_REMOVE))
         {
             if (msg.message == WM_QUIT) {
-                return msg.wParam;  // Received a quit message.
+                return (int)msg.wParam;  // Received a quit message.
             }
 
             TranslateMessage(&msg);
